requirements.txt
openai
pymongo
langchain
python-dotenv
flask
flask_cors
bs4
pinecone-client
confluent-kafka 
